+++ off_chain/model/lotto_composizione_model.py
@@ -10,12 +10,12 @@ from typing import Optional
 class Lotto:
     id_lotto: int
     tipo: str
-    quantita: float
-    cons_co2: float
+    quantita: int
+    cons_co2: int
     composizione: list["Composizione"] = field(default_factory=list)
-    uc_co2 = float
-    co2_costo_composizione : float = 0
-    co2_totale_lotto : float = 0
+    uc_co2 = int
+    co2_costo_composizione : int = 0
+    co2_totale_lotto : int = 0
 
     def __init__(self, id_lotto, tipo, quantita, consumo_co2):
         self.id_lotto = id_lotto
@@ -32,7 +32,7 @@ class Lotto:
         if  self.composizione.__len__:
             for comp in self.composizione:
                 self.co2_costo_composizione += comp.get_co2_consumata_quantità_utilizzata()
-        self.co2_totale_lotto_unitario = (self.co2_costo_composizione + self.cons_co2 ) / self.quantita
+        self.co2_totale_lotto_unitario = int((self.co2_costo_composizione + self.cons_co2 ) / self.quantita)
 
         return self.co2_totale_lotto_unitario 
 
@@ -42,7 +42,7 @@ class Lotto:
 @dataclass
 class Composizione:
     id_lotto_input: int
-    quantita_utilizzata: float
+    quantita_utilizzata: int
     lotto_input: Optional[Lotto] = None
 
     def __init__(self, id_lotto_input, quantita_utilizzata, lotto_input=None):
@@ -50,7 +50,7 @@ class Composizione:
         self.quantita_utilizzata = quantita_utilizzata
         self.lotto_input = lotto_input
 
-    def get_co2_consumata_quantità_utilizzata(self) -> float:
+    def get_co2_consumata_quantità_utilizzata(self) -> int:
         if isinstance(self.lotto_input,Lotto):
             return self.lotto_input.get_costo_totale_lotto_unitario() * self.quantita_utilizzata
         else :
