+++ off_chain/model/operation_estesa_model.py
@@ -16,7 +16,7 @@ class OperazioneEstesaModel:
     id_prodotto : int
     nome_prodotto : str
     quantita_prodotto : int
-    consumo_co2 : float
+    consumo_co2 : int
     blockchain_registered: bool = False  # Indica se l'operazione è registrata sulla blockchain
     descrizione : str = ""  # Campo per la descrizione aggiuntiva del prodotto
